Add files via upload
The main tables for this hospital are to hold doctors,patients and rooms, there is also a table for other staff such as nurses and janitors. Easch tabl has it's own unique primary key with which to identify each entry. Overall a simple description of what one could do with MySQL.